requirements.txt
streamlit
pycryptodome
sphinx
sphinx-autobuild
sphinx_rtd_theme
invoke